Rename Experimenting with Water -Rainbow Code to Rainbow Code for Experiment 3 and 4
This is another slightly changed Rainbow Code I used, I used this code in my third and fourth experiment when I used an old Scentsy cover and a water bottle as a reflective medium. For this I changed the brightness as well as the amount of LED's used. I found that this changed the rate at which the lights changed which I believe benefited my experiment.