Rainbow Code for Experiment 3 and 4
This is another slightly changed Rainbow Code I used, I used this code in my fourth experiment when I used water as a reflective medium. For this I changed the brightness as well as the amount of LED's used. I found that this changed the rate at which the lights changed which I believe benefited my experiment.